+++ tesseraql-maven-plugin/src/main/java/io/tesseraql/maven/MigrateMojo.java
@@ -1,7 +1,7 @@
 package io.tesseraql.maven;
 
 import io.tesseraql.apptasks.AppMigrator;
-import io.tesseraql.report.DriverManagerDataSource;
+import io.tesseraql.core.jdbc.DriverManagerDataSource;
 import io.tesseraql.yaml.manifest.ManifestLoader;
 import io.tesseraql.yaml.migration.SchemaHistoryName;
 import java.io.File;